Fixing fetch for client and server
