+++ package.json
@@ -26,6 +26,8 @@
     "@material-ui/pickers": "^3.2.5",
     "apollo-cache-inmemory": "^1.6.3",
     "apollo-client": "^2.6.4",
+    "apollo-datasource-rest": "^0.6.6",
+    "apollo-link-batch-http": "^1.2.13",
     "apollo-link-http": "^1.5.16",
     "apollo-link-schema": "^1.2.4",
     "apollo-server-express": "^2.9.7",
