+++ src/main/webapp/intrigue-api/fetch.js
@@ -21,6 +21,11 @@ const cacheBust = urlString => {
   const { query, ...rest } = url.parse(urlString)
   return url.format({
     ...rest,
+    hostname: 'localhost',
+    port: '8993',
+    protocol: 'https:',
+    ...(typeof window !== 'undefined' ? window.location : {}),
+    pathname: rest.pathname,
     search: '?' + qs.stringify({ ...qs.parse(query), _: Date.now() }),
   })
 }
@@ -28,7 +33,7 @@ const cacheBust = urlString => {
 module.exports = (url, { headers, ...opts } = {}) => {
   const auth = Buffer.from('admin:admin').toString('base64')
 
-  return fetch(cacheBust('https://localhost:8993' + url), {
+  return fetch(cacheBust(url), {
     credentials: 'same-origin',
     cache: 'no-cache',
     ...opts,
