+++ src/main/webapp/intrigue-api/server.js
@@ -10,13 +10,15 @@ const { resolvers } = require('./graphql')
 const server = new ApolloServer({
   typeDefs,
   resolvers,
-  formatResponse: response => {
-    console.log(response)
-    return response
+  context: ({ req }) => req.headers,
+  formatError: error => {
+    console.log(error)
+    return error
   },
 })
 
 const app = express()
+
 server.applyMiddleware({ app })
 app.listen({ port: 4000 }, () =>
   console.log(`Server ready at http://localhost:4000${server.graphqlPath}`)
